Add log when joined device is loaded from functions Splitted long tests in multiple classes to increase execution paralelism
