+++ LoRaEngine/test/LoRaWan.IntegrationTest/IntegrationTestFixture.cs
@@ -1,3 +1,8 @@
+//
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+//
+
 using System;
 using System.Collections.Generic;
 using System.Linq;
