+++ LoRaEngine/test/LoRaWan.IntegrationTest/LogValidationAssertLevel.cs
@@ -1,3 +1,8 @@
+//
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+//
+
 namespace LoRaWan.IntegrationTest
 {
     // Defines how IoTHub message validation should occur
