+++ LoRaEngine/test/LoRaWanNetworkServer.Test/MessageProcessor_End2End_NoDep_Join_Slow_Get_Twin_Tests.cs
@@ -1,0 +1,135 @@
+﻿//
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+//
+
+using LoRaTools.LoRaMessage;
+using LoRaTools.Utils;
+using LoRaWan.NetworkServer.V2;
+using LoRaWan.Test.Shared;
+using Microsoft.Azure.Devices.Shared;
+using Microsoft.Extensions.Caching.Memory;
+using Moq;
+using System;
+using System.Threading.Tasks;
+using Xunit;
+
+
+namespace LoRaWan.NetworkServer.Test
+{
+    // End to end tests without external dependencies (IoT Hub, Service Facade Function)
+    public class MessageProcessor_End2End_NoDep_Join_Slow_Get_Twin_Tests : MessageProcessorTestBase
+    {
+        [Theory]
+        [InlineData(MessageProcessorTestBase.ServerGatewayID)]
+        [InlineData(null)]
+        public async Task When_Join_Fails_Due_To_Timeout_Second_Try_Should_Reuse_Cached_Device_Twin(string deviceGatewayID)
+        {
+            var simulatedDevice = new SimulatedDevice(TestDeviceInfo.CreateOTAADevice(1, gatewayID: deviceGatewayID));
+            var joinRequest1 = simulatedDevice.CreateJoinRequest();
+
+            // Create Rxpk
+            var joinRequestRxpk1 = joinRequest1.SerializeUplink(simulatedDevice.LoRaDevice.AppKey).rxpk[0];
+
+            var joinRequestDevNonce1 = LoRaTools.Utils.ConversionHelper.ByteArrayToString(joinRequest1.DevNonce);
+            var devAddr = string.Empty;
+            var devEUI = simulatedDevice.LoRaDevice.DeviceID;
+            var appEUI = simulatedDevice.LoRaDevice.AppEUI;
+
+            var loRaDeviceClient = new Mock<ILoRaDeviceClient>(MockBehavior.Strict);
+
+            // Device twin will be queried twice, 1st time will take 7 seconds, 2nd time 0.1 second
+            var twin = new Twin();
+            twin.Properties.Desired[TwinProperty.DevEUI] = devEUI;
+            twin.Properties.Desired[TwinProperty.AppEUI] = simulatedDevice.LoRaDevice.AppEUI;
+            twin.Properties.Desired[TwinProperty.AppKey] = simulatedDevice.LoRaDevice.AppKey;
+            twin.Properties.Desired[TwinProperty.GatewayID] = deviceGatewayID;
+            twin.Properties.Desired[TwinProperty.SensorDecoder] = simulatedDevice.LoRaDevice.SensorDecoder;
+            loRaDeviceClient.SetupSequence(x => x.GetTwinAsync())
+                .Returns(async () =>
+                {
+                    await Task.Delay(TimeSpan.FromSeconds(7));
+                    return twin;
+                })
+                .Returns(async () =>
+                {
+                    await Task.Delay(TimeSpan.FromSeconds(0.1));
+                    return twin;
+                });
+
+            // Device twin will be updated
+            string afterJoinAppSKey = null;
+            string afterJoinNwkSKey = null;
+            string afterJoinDevAddr = null;
+            loRaDeviceClient.Setup(x => x.UpdateReportedPropertiesAsync(It.IsNotNull<TwinCollection>()))
+                .Callback<TwinCollection>((updatedTwin) =>
+                {
+                    afterJoinAppSKey = updatedTwin[TwinProperty.AppSKey];
+                    afterJoinNwkSKey = updatedTwin[TwinProperty.NwkSKey];
+                    afterJoinDevAddr = updatedTwin[TwinProperty.DevAddr];
+                })
+                .ReturnsAsync(true);
+
+            // Lora device api will be search by devices with matching deveui,
+            var loRaDeviceApi = new Mock<LoRaDeviceAPIServiceBase>(MockBehavior.Strict);
+            loRaDeviceApi.Setup(x => x.SearchDevicesAsync(ServerConfiguration.GatewayID, null, devEUI, appEUI, joinRequestDevNonce1))
+                .ReturnsAsync(new SearchDevicesResult(new IoTHubDeviceInfo(devAddr, devEUI, "aabb").AsList()));
+
+            // using factory to create mock of 
+            var loRaDeviceFactory = new TestLoRaDeviceFactory(loRaDeviceClient.Object);
+
+            var memoryCache = new MemoryCache(new MemoryCacheOptions());
+            var deviceRegistry = new LoRaDeviceRegistry(this.ServerConfiguration, memoryCache, loRaDeviceApi.Object, loRaDeviceFactory);
+
+            var frameCounterUpdateStrategyFactory = new LoRaDeviceFrameCounterUpdateStrategyFactory(ServerConfiguration.GatewayID, loRaDeviceApi.Object);
+
+            var messageProcessor = new LoRaWan.NetworkServer.V2.MessageProcessor(
+                this.ServerConfiguration,
+                deviceRegistry,
+                frameCounterUpdateStrategyFactory,
+                new LoRaPayloadDecoder()
+                );
+
+            // 1st join request
+            // Should fail
+            var joinRequestDownlinkMessage1 = await messageProcessor.ProcessMessageAsync(joinRequestRxpk1);
+            Assert.Null(joinRequestDownlinkMessage1);
+
+            // 2nd attempt
+            var joinRequest2 = simulatedDevice.CreateJoinRequest();
+            var joinRequestRxpk2 = joinRequest2.SerializeUplink(simulatedDevice.LoRaDevice.AppKey).rxpk[0];
+            var joinRequestDevNonce2 = LoRaTools.Utils.ConversionHelper.ByteArrayToString(joinRequest2.DevNonce);
+
+            // setup response to this device search
+            loRaDeviceApi.Setup(x => x.SearchDevicesAsync(ServerConfiguration.GatewayID, null, devEUI, appEUI, joinRequestDevNonce2))
+                .ReturnsAsync(new SearchDevicesResult(new IoTHubDeviceInfo(devAddr, devEUI, "aabb").AsList()));
+
+            var joinRequestDownlinkMessage2 = await messageProcessor.ProcessMessageAsync(joinRequestRxpk2);
+            Assert.NotNull(joinRequestDownlinkMessage2);
+            var joinAccept = new LoRaPayloadJoinAccept(Convert.FromBase64String(joinRequestDownlinkMessage2.txpk.data), simulatedDevice.LoRaDevice.AppKey);
+            Assert.Equal(joinAccept.DevAddr.ToArray(), ConversionHelper.StringToByteArray(afterJoinDevAddr));
+
+            var devicesForDevAddr = deviceRegistry.InternalGetCachedDevicesForDevAddr(afterJoinDevAddr);
+            Assert.Single(devicesForDevAddr); // should have the single device
+            Assert.True(devicesForDevAddr.TryGetValue(devEUI, out var loRaDevice));
+
+            Assert.Equal(simulatedDevice.AppKey, loRaDevice.AppKey);
+            Assert.Equal(simulatedDevice.AppEUI, loRaDevice.AppEUI);
+            Assert.Equal(afterJoinAppSKey, loRaDevice.AppSKey);
+            Assert.Equal(afterJoinNwkSKey, loRaDevice.NwkSKey);
+            Assert.Equal(afterJoinDevAddr, loRaDevice.DevAddr);
+            if (deviceGatewayID == null)
+                Assert.Null(loRaDevice.GatewayID);
+            else
+                Assert.Equal(deviceGatewayID, loRaDevice.GatewayID);
+
+            // fcnt is restarted
+            Assert.Equal(0, loRaDevice.FCntUp);
+            Assert.Equal(0, loRaDevice.FCntDown);
+            Assert.False(loRaDevice.HasFrameCountChanges);
+
+            loRaDeviceClient.VerifyAll();
+            loRaDeviceApi.VerifyAll();
+        }
+    }
+}
