+++ LoRaEngine/test/LoRaWanNetworkServer.Test/MessageProcessor_End2End_NoDep_Processing_Tests.cs
@@ -1,4 +1,10 @@
-﻿using LoRaTools.LoRaMessage;
+﻿//
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+//
+
+using LoRaTools.LoRaMessage;
+using LoRaTools.LoRaPhysical;
 using LoRaTools.Regions;
 using LoRaTools.Utils;
 using LoRaWan.NetworkServer.V2;
@@ -7,6 +13,7 @@ using Microsoft.Azure.Devices.Client;
 using Microsoft.Azure.Devices.Shared;
 using Microsoft.Extensions.Caching.Memory;
 using Moq;
+using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
@@ -15,174 +22,20 @@ using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
 
+
 namespace LoRaWan.NetworkServer.Test
 {
     // End to end tests without external dependencies (IoT Hub, Service Facade Function)
-    public class MessageProcessor_End2End_NoDep_Tests : MessageProcessorTestBase
+    // General message processor tests (Join tests are handled in other class)
+    public class MessageProcessor_End2End_NoDep_Processing_Tests : MessageProcessorTestBase
     {
        
-        public MessageProcessor_End2End_NoDep_Tests()
+        public MessageProcessor_End2End_NoDep_Processing_Tests()
         {
           
         }
 
-
-        [Theory]
-        [InlineData(MessageProcessorTestBase.ServerGatewayID, 200, 50)]
-        [InlineData(MessageProcessorTestBase.ServerGatewayID, 0, 0)]
-        [InlineData(null, 200, 50)]
-        [InlineData(null, 0, 0)]
-        public async Task Join_And_Send_Unconfirmed_And_Confirmed_Messages(string deviceGatewayID, int initialFcntUp, int initialFcntDown)
-        {         
-            var simulatedDevice = new SimulatedDevice(TestDeviceInfo.CreateOTAADevice(1, gatewayID: deviceGatewayID));
-            var joinRequest = simulatedDevice.CreateJoinRequest();
-
-            // Create Rxpk
-            var joinRxpk = joinRequest.SerializeUplink(simulatedDevice.LoRaDevice.AppKey).rxpk[0];
-            
-
-            var devNonce = LoRaTools.Utils.ConversionHelper.ByteArrayToString(joinRequest.DevNonce);
-            var devAddr = string.Empty;
-            var devEUI = simulatedDevice.LoRaDevice.DeviceID;
-            var appEUI = simulatedDevice.LoRaDevice.AppEUI;
-
-            var loRaDeviceClient = new Mock<ILoRaDeviceClient>(MockBehavior.Strict);
-
-            // Device twin will be queried
-            var twin = new Twin();
-            twin.Properties.Desired[TwinProperty.DevEUI] = devEUI;
-            twin.Properties.Desired[TwinProperty.AppEUI] = simulatedDevice.LoRaDevice.AppEUI;
-            twin.Properties.Desired[TwinProperty.AppKey] = simulatedDevice.LoRaDevice.AppKey;
-            twin.Properties.Desired[TwinProperty.GatewayID] = deviceGatewayID;
-            twin.Properties.Desired[TwinProperty.SensorDecoder] = simulatedDevice.LoRaDevice.SensorDecoder;
-            twin.Properties.Reported[TwinProperty.FCntUp] = initialFcntUp;
-            twin.Properties.Reported[TwinProperty.FCntDown] = initialFcntDown;
-            loRaDeviceClient.Setup(x => x.GetTwinAsync()).ReturnsAsync(twin);
-
-            // Device twin will be updated
-            string afterJoinAppSKey = null;
-            string afterJoinNwkSKey = null;
-            string afterJoinDevAddr = null;
-            int afterJoinFcntDown = -1;
-            int afterJoinFcntUp = -1;
-            loRaDeviceClient.Setup(x => x.UpdateReportedPropertiesAsync(It.IsNotNull<TwinCollection>()))
-                .Callback<TwinCollection>((updatedTwin) => {
-                    afterJoinAppSKey = updatedTwin[TwinProperty.AppSKey];
-                    afterJoinNwkSKey = updatedTwin[TwinProperty.NwkSKey];
-                    afterJoinDevAddr = updatedTwin[TwinProperty.DevAddr];
-                    afterJoinFcntDown = updatedTwin[TwinProperty.FCntDown];
-                    afterJoinFcntUp = updatedTwin[TwinProperty.FCntUp];
-                })
-                .ReturnsAsync(true);
-
-            // message will be sent
-            loRaDeviceClient.Setup(x => x.SendEventAsync(It.IsNotNull<LoRaDeviceTelemetry>(), null))
-                .ReturnsAsync(true);
-
-            // C2D message will be checked
-            loRaDeviceClient.Setup(x => x.ReceiveAsync(It.IsNotNull<TimeSpan>()))
-                .ReturnsAsync((Message)null);
-
-            // Lora device api will be search by devices with matching deveui,
-            var loRaDeviceApi = new Mock<LoRaDeviceAPIServiceBase>(MockBehavior.Strict);
-            loRaDeviceApi.Setup(x => x.SearchDevicesAsync(ServerConfiguration.GatewayID, null, devEUI, appEUI, devNonce))
-                .ReturnsAsync(new SearchDevicesResult(new IoTHubDeviceInfo(devAddr, devEUI, "aabb").AsList()));            
-
-
-            // multi gateway will request a next frame count down from the lora device api, prepare it
-            if (string.IsNullOrEmpty(deviceGatewayID))
-            {
-                loRaDeviceApi.Setup(x => x.NextFCntDownAsync(devEUI, 0, 2, ServerConfiguration.GatewayID))
-                    .ReturnsAsync((ushort)1);
-            }
-
-            // using factory to create mock of 
-            var loRaDeviceFactory = new TestLoRaDeviceFactory(loRaDeviceClient.Object);
-
-            var memoryCache = new MemoryCache(new MemoryCacheOptions());
-            var deviceRegistry = new LoRaDeviceRegistry(this.ServerConfiguration, memoryCache, loRaDeviceApi.Object, loRaDeviceFactory);
-
-            var frameCounterUpdateStrategyFactory = new LoRaDeviceFrameCounterUpdateStrategyFactory(ServerConfiguration.GatewayID, loRaDeviceApi.Object);
-            
-
-            // Send to message processor
-            var messageProcessor = new LoRaWan.NetworkServer.V2.MessageProcessor(
-                this.ServerConfiguration,
-                deviceRegistry,
-                frameCounterUpdateStrategyFactory,
-                new LoRaPayloadDecoder()
-                );
-
-            var downlinkJoinAcceptMessage = await messageProcessor.ProcessMessageAsync(joinRxpk);
-            Assert.NotNull(downlinkJoinAcceptMessage);
-            var joinAccept = new LoRaPayloadJoinAccept(Convert.FromBase64String(downlinkJoinAcceptMessage.txpk.data), simulatedDevice.LoRaDevice.AppKey);
-            Assert.Equal(joinAccept.DevAddr.ToArray(), ConversionHelper.StringToByteArray(afterJoinDevAddr));
-
-            // check that the device is in cache
-            var devicesForDevAddr = deviceRegistry.InternalGetCachedDevicesForDevAddr(afterJoinDevAddr);
-            Assert.Single(devicesForDevAddr); // should have the single device
-            Assert.True(devicesForDevAddr.TryGetValue(devEUI, out var loRaDevice));
-            Assert.Equal(afterJoinAppSKey, loRaDevice.AppSKey);
-            Assert.Equal(afterJoinNwkSKey, loRaDevice.NwkSKey);
-            Assert.Equal(afterJoinDevAddr, loRaDevice.DevAddr);
-
-            if (deviceGatewayID == null)
-                Assert.Null(loRaDevice.GatewayID);
-            else
-                Assert.Equal(deviceGatewayID, loRaDevice.GatewayID);
-
-            // fcnt is restarted
-            Assert.Equal(0, afterJoinFcntDown);
-            Assert.Equal(0, afterJoinFcntUp);
-            Assert.Equal(0, loRaDevice.FCntUp);
-            Assert.Equal(0, loRaDevice.FCntDown);
-            Assert.False(loRaDevice.HasFrameCountChanges);
-
-
-            simulatedDevice.LoRaDevice.AppSKey = afterJoinAppSKey;
-            simulatedDevice.LoRaDevice.NwkSKey = afterJoinNwkSKey;
-            simulatedDevice.LoRaDevice.DevAddr = afterJoinDevAddr;
-
-            // sends unconfirmed message
-            var unconfirmedMessagePayload = simulatedDevice.CreateUnconfirmedDataUpMessage("100", fcnt: 1);
-            var unconfirmedMessageResult = await messageProcessor.ProcessMessageAsync(unconfirmedMessagePayload.SerializeUplink(simulatedDevice.AppSKey, simulatedDevice.NwkSKey).rxpk[0]);
-            Assert.Null(unconfirmedMessageResult);
-
-            // fcnt up was updated
-            Assert.Equal(1, loRaDevice.FCntUp);
-            Assert.Equal(0, loRaDevice.FCntDown);
-
-            // Frame change flag will be set, only saving every 10 messages
-            Assert.True(loRaDevice.HasFrameCountChanges);
-
-
-            // sends confirmed message
-            var confirmedMessagePayload = simulatedDevice.CreateConfirmedDataUpMessage("200", fcnt: 2);
-            var confirmedMessageRxpk = confirmedMessagePayload.SerializeUplink(simulatedDevice.AppSKey, simulatedDevice.NwkSKey).rxpk[0];
-            var confirmedMessage = await messageProcessor.ProcessMessageAsync(confirmedMessageRxpk);
-            Assert.NotNull(confirmedMessage);
-            Assert.NotNull(confirmedMessage.txpk);
-           
-
-            // validates txpk according to eu region
-            Assert.Equal(RegionFactory.CreateEU868Region().GetDownstreamChannel(confirmedMessageRxpk), confirmedMessage.txpk.freq);
-            Assert.Equal("4/5", confirmedMessage.txpk.codr);
-            Assert.False(confirmedMessage.txpk.imme);
-            Assert.True(confirmedMessage.txpk.ipol);
-            Assert.Equal("LORA", confirmedMessage.txpk.modu);
-
-            // fcnt up was updated
-            Assert.Equal(2, loRaDevice.FCntUp);
-            Assert.Equal(1, loRaDevice.FCntDown);
-
-            // Frame change flag will be set, only saving every 10 messages
-            Assert.True(loRaDevice.HasFrameCountChanges);
-
-            loRaDeviceClient.VerifyAll();
-            loRaDeviceApi.VerifyAll();
-        }
-
-        [Theory]
+       [Theory]
         [InlineData(null, 0, null, null)]
         [InlineData(null, 0, 1, 1)]
         [InlineData(null, 0, 100, 20)]
@@ -683,452 +536,6 @@ namespace LoRaWan.NetworkServer.Test
             loRaDeviceClient.VerifyAll();
         }
 
-
-        [Theory]
-        [InlineData(MessageProcessorTestBase.ServerGatewayID)]
-        [InlineData(null)]
-        public async Task When_Join_Fails_Due_To_Timeout_Second_Try_Should_Reuse_Cached_Device_Twin(string deviceGatewayID)
-        {
-            var simulatedDevice = new SimulatedDevice(TestDeviceInfo.CreateOTAADevice(1, gatewayID: deviceGatewayID));
-            var joinRequest1 = simulatedDevice.CreateJoinRequest();
-
-            // Create Rxpk
-            var joinRequestRxpk1 = joinRequest1.SerializeUplink(simulatedDevice.LoRaDevice.AppKey).rxpk[0];
-
-            var joinRequestDevNonce1 = LoRaTools.Utils.ConversionHelper.ByteArrayToString(joinRequest1.DevNonce);
-            var devAddr = string.Empty;
-            var devEUI = simulatedDevice.LoRaDevice.DeviceID;
-            var appEUI = simulatedDevice.LoRaDevice.AppEUI;
-
-            var loRaDeviceClient = new Mock<ILoRaDeviceClient>(MockBehavior.Strict);
-
-            // Device twin will be queried twice, 1st time will take 7 seconds, 2nd time 0.1 second
-            var twin = new Twin();
-            twin.Properties.Desired[TwinProperty.DevEUI] = devEUI;
-            twin.Properties.Desired[TwinProperty.AppEUI] = simulatedDevice.LoRaDevice.AppEUI;
-            twin.Properties.Desired[TwinProperty.AppKey] = simulatedDevice.LoRaDevice.AppKey;
-            twin.Properties.Desired[TwinProperty.GatewayID] = deviceGatewayID;
-            twin.Properties.Desired[TwinProperty.SensorDecoder] = simulatedDevice.LoRaDevice.SensorDecoder;
-            loRaDeviceClient.SetupSequence(x => x.GetTwinAsync())
-                .Returns(async () =>
-                {
-                    await Task.Delay(TimeSpan.FromSeconds(7));
-                    return twin;
-                })
-                .Returns(async () =>
-                {
-                    await Task.Delay(TimeSpan.FromSeconds(0.1));
-                    return twin;
-                });
-
-            // Device twin will be updated
-            string afterJoinAppSKey = null;
-            string afterJoinNwkSKey = null;
-            string afterJoinDevAddr = null;
-            loRaDeviceClient.Setup(x => x.UpdateReportedPropertiesAsync(It.IsNotNull<TwinCollection>()))
-                .Callback<TwinCollection>((updatedTwin) => {
-                    afterJoinAppSKey = updatedTwin[TwinProperty.AppSKey];
-                    afterJoinNwkSKey = updatedTwin[TwinProperty.NwkSKey];
-                    afterJoinDevAddr = updatedTwin[TwinProperty.DevAddr];
-                })
-                .ReturnsAsync(true);           
-
-            // Lora device api will be search by devices with matching deveui,
-            var loRaDeviceApi = new Mock<LoRaDeviceAPIServiceBase>(MockBehavior.Strict);
-            loRaDeviceApi.Setup(x => x.SearchDevicesAsync(ServerConfiguration.GatewayID, null, devEUI, appEUI, joinRequestDevNonce1))
-                .ReturnsAsync(new SearchDevicesResult(new IoTHubDeviceInfo(devAddr, devEUI, "aabb").AsList()));
-
-            // using factory to create mock of 
-            var loRaDeviceFactory = new TestLoRaDeviceFactory(loRaDeviceClient.Object);
-
-            var memoryCache = new MemoryCache(new MemoryCacheOptions());
-            var deviceRegistry = new LoRaDeviceRegistry(this.ServerConfiguration, memoryCache, loRaDeviceApi.Object, loRaDeviceFactory);
-
-            var frameCounterUpdateStrategyFactory = new LoRaDeviceFrameCounterUpdateStrategyFactory(ServerConfiguration.GatewayID, loRaDeviceApi.Object);
-            
-            var messageProcessor = new LoRaWan.NetworkServer.V2.MessageProcessor(
-                this.ServerConfiguration,
-                deviceRegistry,
-                frameCounterUpdateStrategyFactory,
-                new LoRaPayloadDecoder()
-                );
-
-            // 1st join request
-            // Should fail
-            var joinRequestDownlinkMessage1 = await messageProcessor.ProcessMessageAsync(joinRequestRxpk1);
-            Assert.Null(joinRequestDownlinkMessage1);
-
-            // 2nd attempt
-            var joinRequest2 = simulatedDevice.CreateJoinRequest();
-            var joinRequestRxpk2 = joinRequest2.SerializeUplink(simulatedDevice.LoRaDevice.AppKey).rxpk[0];
-            var joinRequestDevNonce2 = LoRaTools.Utils.ConversionHelper.ByteArrayToString(joinRequest2.DevNonce);
-
-            // setup response to this device search
-            loRaDeviceApi.Setup(x => x.SearchDevicesAsync(ServerConfiguration.GatewayID, null, devEUI, appEUI, joinRequestDevNonce2))
-                .ReturnsAsync(new SearchDevicesResult(new IoTHubDeviceInfo(devAddr, devEUI, "aabb").AsList()));
-
-            var joinRequestDownlinkMessage2 = await messageProcessor.ProcessMessageAsync(joinRequestRxpk2);
-            Assert.NotNull(joinRequestDownlinkMessage2);
-            var joinAccept = new LoRaPayloadJoinAccept(Convert.FromBase64String(joinRequestDownlinkMessage2.txpk.data), simulatedDevice.LoRaDevice.AppKey);
-            Assert.Equal(joinAccept.DevAddr.ToArray(), ConversionHelper.StringToByteArray(afterJoinDevAddr));
-
-            var devicesForDevAddr = deviceRegistry.InternalGetCachedDevicesForDevAddr(afterJoinDevAddr);
-            Assert.Single(devicesForDevAddr); // should have the single device
-            Assert.True(devicesForDevAddr.TryGetValue(devEUI, out var loRaDevice));
-
-            Assert.Equal(simulatedDevice.AppKey, loRaDevice.AppKey);
-            Assert.Equal(simulatedDevice.AppEUI, loRaDevice.AppEUI);
-            Assert.Equal(afterJoinAppSKey, loRaDevice.AppSKey);
-            Assert.Equal(afterJoinNwkSKey, loRaDevice.NwkSKey);
-            Assert.Equal(afterJoinDevAddr, loRaDevice.DevAddr);
-            if (deviceGatewayID == null)
-                Assert.Null(loRaDevice.GatewayID);
-            else
-                Assert.Equal(deviceGatewayID, loRaDevice.GatewayID);
-
-            // fcnt is restarted
-            Assert.Equal(0, loRaDevice.FCntUp);
-            Assert.Equal(0, loRaDevice.FCntDown);
-            Assert.False(loRaDevice.HasFrameCountChanges);
-
-            loRaDeviceClient.VerifyAll();
-            loRaDeviceApi.VerifyAll();
-        }
-
-
-
-        [Theory]
-        [InlineData(MessageProcessorTestBase.ServerGatewayID)]
-        [InlineData(null)]
-        public async Task When_Join_Fails_Due_To_GetTwin_Error_Second_Try_Should_Reload_Device_Twin(string deviceGatewayID)
-        {
-            var simulatedDevice = new SimulatedDevice(TestDeviceInfo.CreateOTAADevice(1, gatewayID: deviceGatewayID));
-            var joinRequest1 = simulatedDevice.CreateJoinRequest();
-
-            // Create Rxpk
-            var joinRequestRxpk1 = joinRequest1.SerializeUplink(simulatedDevice.LoRaDevice.AppKey).rxpk[0];
-
-            var joinRequestDevNonce1 = ConversionHelper.ByteArrayToString(joinRequest1.DevNonce);
-            var devAddr = string.Empty;
-            var devEUI = simulatedDevice.LoRaDevice.DeviceID;
-            var appEUI = simulatedDevice.LoRaDevice.AppEUI;
-
-            var loRaDeviceClient = new Mock<ILoRaDeviceClient>(MockBehavior.Strict);
-
-            // Device twin will be queried
-            var twin = new Twin();
-            twin.Properties.Desired[TwinProperty.DevEUI] = devEUI;
-            twin.Properties.Desired[TwinProperty.AppEUI] = simulatedDevice.LoRaDevice.AppEUI;
-            twin.Properties.Desired[TwinProperty.AppKey] = simulatedDevice.LoRaDevice.AppKey;
-            twin.Properties.Desired[TwinProperty.GatewayID] = deviceGatewayID;
-            twin.Properties.Desired[TwinProperty.SensorDecoder] = simulatedDevice.LoRaDevice.SensorDecoder;
-            loRaDeviceClient.SetupSequence(x => x.GetTwinAsync())
-                .ReturnsAsync((Twin)null)
-                .ReturnsAsync(twin);
-
-            // Device twin will be updated
-            string afterJoinAppSKey = null;
-            string afterJoinNwkSKey = null;
-            string afterJoinDevAddr = null;
-            loRaDeviceClient.Setup(x => x.UpdateReportedPropertiesAsync(It.IsNotNull<TwinCollection>()))
-                .Callback<TwinCollection>((updatedTwin) => {
-                    afterJoinAppSKey = updatedTwin[TwinProperty.AppSKey];
-                    afterJoinNwkSKey = updatedTwin[TwinProperty.NwkSKey];
-                    afterJoinDevAddr = updatedTwin[TwinProperty.DevAddr];
-                })
-                .ReturnsAsync(true);
-
-            // Lora device api will be search by devices with matching deveui,
-            var loRaDeviceApi = new Mock<LoRaDeviceAPIServiceBase>(MockBehavior.Strict);
-            loRaDeviceApi.Setup(x => x.SearchDevicesAsync(ServerConfiguration.GatewayID, null, devEUI, appEUI, joinRequestDevNonce1))
-                .ReturnsAsync(new SearchDevicesResult(new IoTHubDeviceInfo(devAddr, devEUI, "aabb").AsList()));
-
-            // using factory to create mock of 
-            var loRaDeviceFactory = new TestLoRaDeviceFactory(loRaDeviceClient.Object);
-
-            var memoryCache = new MemoryCache(new MemoryCacheOptions());
-            var deviceRegistry = new LoRaDeviceRegistry(this.ServerConfiguration, memoryCache, loRaDeviceApi.Object, loRaDeviceFactory);
-
-            var frameCounterUpdateStrategyFactory = new LoRaDeviceFrameCounterUpdateStrategyFactory(ServerConfiguration.GatewayID, loRaDeviceApi.Object);
-
-            var messageProcessor = new LoRaWan.NetworkServer.V2.MessageProcessor(
-                this.ServerConfiguration,
-                deviceRegistry,
-                frameCounterUpdateStrategyFactory,
-                new LoRaPayloadDecoder()
-                );
-
-            // 1st join request
-            // Should fail
-            var joinRequestDownlinkMessage1 = await messageProcessor.ProcessMessageAsync(joinRequestRxpk1);
-            Assert.Null(joinRequestDownlinkMessage1);
-
-            // 2nd attempt
-            var joinRequest2 = simulatedDevice.CreateJoinRequest();
-
-            // will reload the device matched by deveui
-            loRaDeviceApi.Setup(x => x.SearchDevicesAsync(ServerConfiguration.GatewayID, null, devEUI, appEUI, ConversionHelper.ByteArrayToString(joinRequest2.DevNonce)))
-                .ReturnsAsync(new SearchDevicesResult(new IoTHubDeviceInfo(devAddr, devEUI, "aabb").AsList()));
-
-            var joinRequestRxpk2 = joinRequest2.SerializeUplink(simulatedDevice.LoRaDevice.AppKey).rxpk[0];
-            var joinRequestDevNonce2 = LoRaTools.Utils.ConversionHelper.ByteArrayToString(joinRequest2.DevNonce);
-            var joinRequestDownlinkMessage2 = await messageProcessor.ProcessMessageAsync(joinRequestRxpk2);
-            Assert.NotNull(joinRequestDownlinkMessage2);
-            var joinAccept = new LoRaPayloadJoinAccept(Convert.FromBase64String(joinRequestDownlinkMessage2.txpk.data), simulatedDevice.LoRaDevice.AppKey);
-            Assert.Equal(joinAccept.DevAddr.ToArray(), ConversionHelper.StringToByteArray(afterJoinDevAddr));
-
-            var devicesForDevAddr = deviceRegistry.InternalGetCachedDevicesForDevAddr(afterJoinDevAddr);
-            Assert.Single(devicesForDevAddr); // should have the single device
-            Assert.True(devicesForDevAddr.TryGetValue(devEUI, out var loRaDevice));
-
-            Assert.Equal(simulatedDevice.AppKey, loRaDevice.AppKey);
-            Assert.Equal(simulatedDevice.AppEUI, loRaDevice.AppEUI);
-            Assert.Equal(afterJoinAppSKey, loRaDevice.AppSKey);
-            Assert.Equal(afterJoinNwkSKey, loRaDevice.NwkSKey);
-            Assert.Equal(afterJoinDevAddr, loRaDevice.DevAddr);
-            if (deviceGatewayID == null)
-                Assert.Null(loRaDevice.GatewayID);
-            else
-                Assert.Equal(deviceGatewayID, loRaDevice.GatewayID);
-
-            // fcnt is restarted
-            Assert.Equal(0, loRaDevice.FCntUp);
-            Assert.Equal(0, loRaDevice.FCntDown);
-            Assert.False(loRaDevice.HasFrameCountChanges);
-
-            loRaDeviceClient.VerifyAll();
-            loRaDeviceApi.VerifyAll();
-        }
-
-        /// <summary>
-        /// Verifies that if the update twin takes too long that no join accepts are sent
-        /// </summary>
-        /// <param name="deviceGatewayID"></param>
-        /// <returns></returns>
-        [Theory]
-        [InlineData(MessageProcessorTestBase.ServerGatewayID)]
-        [InlineData(null)]
-        public async Task When_First_Join_Fails_Due_To_Slow_Twin_Update_Retry_Second_Attempt_Should_Succeed(string deviceGatewayID)
-        {
-            var simulatedDevice = new SimulatedDevice(TestDeviceInfo.CreateOTAADevice(1, gatewayID: deviceGatewayID));
-            var joinRequest1 = simulatedDevice.CreateJoinRequest();
-            var joinRequest2 = simulatedDevice.CreateJoinRequest();
-
-            // Create Rxpk
-            var joinRequestRxpk1 = joinRequest1.SerializeUplink(simulatedDevice.LoRaDevice.AppKey).rxpk[0];
-            var joinRequestRxpk2 = joinRequest2.SerializeUplink(simulatedDevice.LoRaDevice.AppKey).rxpk[0];
-
-
-            var joinRequestDevNonce1 = LoRaTools.Utils.ConversionHelper.ByteArrayToString(joinRequest1.DevNonce);
-            var joinRequestDevNonce2 = LoRaTools.Utils.ConversionHelper.ByteArrayToString(joinRequest2.DevNonce);
-            var devAddr = string.Empty;
-            var devEUI = simulatedDevice.LoRaDevice.DeviceID;
-            var appEUI = simulatedDevice.LoRaDevice.AppEUI;
-
-            var loRaDeviceClient = new Mock<ILoRaDeviceClient>(MockBehavior.Strict);
-
-            // Device twin will be queried
-            var twin = new Twin();
-            twin.Properties.Desired[TwinProperty.DevEUI] = devEUI;
-            twin.Properties.Desired[TwinProperty.AppEUI] = simulatedDevice.LoRaDevice.AppEUI;
-            twin.Properties.Desired[TwinProperty.AppKey] = simulatedDevice.LoRaDevice.AppKey;
-            twin.Properties.Desired[TwinProperty.GatewayID] = deviceGatewayID;
-            twin.Properties.Desired[TwinProperty.SensorDecoder] = simulatedDevice.LoRaDevice.SensorDecoder;
-            loRaDeviceClient.Setup(x => x.GetTwinAsync())
-                .ReturnsAsync(twin);
-
-            // Device twin will be updated
-            string afterJoin1AppSKey = null;
-            string afterJoin1NwkSKey = null;
-            string afterJoin1DevAddr = null;
-            string afterJoin2AppSKey = null;
-            string afterJoin2NwkSKey = null;
-            string afterJoin2DevAddr = null;
-            var isFirstTwinUpdate = true;
-            loRaDeviceClient.Setup(x => x.UpdateReportedPropertiesAsync(It.IsNotNull<TwinCollection>()))
-                .Callback<TwinCollection>((updatedTwin) => {
-                    if (isFirstTwinUpdate)
-                    {
-                        afterJoin1AppSKey = updatedTwin[TwinProperty.AppSKey];
-                        afterJoin1NwkSKey = updatedTwin[TwinProperty.NwkSKey];
-                        afterJoin1DevAddr = updatedTwin[TwinProperty.DevAddr];
-
-                        Thread.Sleep(TimeSpan.FromSeconds(10));
-                        isFirstTwinUpdate = false;
-                    }
-                    else
-                    {
-                        afterJoin2AppSKey = updatedTwin[TwinProperty.AppSKey];
-                        afterJoin2NwkSKey = updatedTwin[TwinProperty.NwkSKey];
-                        afterJoin2DevAddr = updatedTwin[TwinProperty.DevAddr];
-                    }
-                })
-                .ReturnsAsync(true);
-
-            // Lora device api will be search by devices with matching deveui,
-            var loRaDeviceApi = new Mock<LoRaDeviceAPIServiceBase>(MockBehavior.Strict);
-            loRaDeviceApi.Setup(x => x.SearchDevicesAsync(ServerConfiguration.GatewayID, null, devEUI, appEUI, joinRequestDevNonce1))
-                .ReturnsAsync(new SearchDevicesResult(new IoTHubDeviceInfo(devAddr, devEUI, "aabb").AsList()));
-
-            loRaDeviceApi.Setup(x => x.SearchDevicesAsync(ServerConfiguration.GatewayID, null, devEUI, appEUI, joinRequestDevNonce2))
-                .ReturnsAsync(new SearchDevicesResult(new IoTHubDeviceInfo(devAddr, devEUI, "aabb").AsList()));
-
-            // using factory to create mock of 
-            var loRaDeviceFactory = new TestLoRaDeviceFactory(loRaDeviceClient.Object);
-
-            var memoryCache = new MemoryCache(new MemoryCacheOptions());
-            var deviceRegistry = new LoRaDeviceRegistry(this.ServerConfiguration, memoryCache, loRaDeviceApi.Object, loRaDeviceFactory);
-
-            var frameCounterUpdateStrategyFactory = new LoRaDeviceFrameCounterUpdateStrategyFactory(ServerConfiguration.GatewayID, loRaDeviceApi.Object);
-
-            var messageProcessor = new LoRaWan.NetworkServer.V2.MessageProcessor(
-                this.ServerConfiguration,
-                deviceRegistry,
-                frameCounterUpdateStrategyFactory,
-                new LoRaPayloadDecoder()
-                );
-
-            var joinRequestTask1 = messageProcessor.ProcessMessageAsync(joinRequestRxpk1);
-            
-            await Task.Delay(TimeSpan.FromSeconds(7));
-
-            var joinRequestTask2 = messageProcessor.ProcessMessageAsync(joinRequestRxpk2);
-
-            await Task.WhenAll(joinRequestTask1, joinRequestTask2);                       
-            Assert.Null(joinRequestTask1.Result);
-            Assert.NotNull(joinRequestTask2.Result);
-
-            Assert.Empty(deviceRegistry.InternalGetCachedDevicesForDevAddr(afterJoin1DevAddr));
-            var devicesInDevAddr2 = deviceRegistry.InternalGetCachedDevicesForDevAddr(afterJoin2DevAddr);
-            Assert.NotEmpty(devicesInDevAddr2);
-            Assert.True(devicesInDevAddr2.TryGetValue(devEUI, out var loRaDevice));
-            Assert.True(loRaDevice.IsOurDevice);
-            Assert.Equal(afterJoin2DevAddr, loRaDevice.DevAddr);
-            Assert.Equal(afterJoin2NwkSKey, loRaDevice.NwkSKey);
-            Assert.Equal(afterJoin2AppSKey, loRaDevice.AppSKey);
-
-            loRaDeviceClient.VerifyAll();
-            loRaDeviceApi.VerifyAll();
-        }
-
-
-
-
-        [Theory]
-        [InlineData(MessageProcessorTestBase.ServerGatewayID)]
-        [InlineData(null)]
-        public async Task Join_Device_Has_Mismatching_AppEUI_Should_Return_Null(string deviceGatewayID)
-        {
-            var simulatedDevice = new SimulatedDevice(TestDeviceInfo.CreateOTAADevice(1, gatewayID: deviceGatewayID));
-            var joinRequest = simulatedDevice.CreateJoinRequest();
-
-            // Create Rxpk
-            var joinRequestRxpk = joinRequest.SerializeUplink(simulatedDevice.AppKey).rxpk[0];
-
-            var joinRequestDevNonce = LoRaTools.Utils.ConversionHelper.ByteArrayToString(joinRequest.DevNonce);
-            var devAddr = string.Empty;
-            var devEUI = simulatedDevice.LoRaDevice.DeviceID;
-            var appEUI = simulatedDevice.LoRaDevice.AppEUI;
-
-            var loRaDeviceClient = new Mock<ILoRaDeviceClient>(MockBehavior.Strict);
-
-            // Device twin will be queried
-            var twin = new Twin();
-            twin.Properties.Desired[TwinProperty.DevEUI] = devEUI;
-            twin.Properties.Desired[TwinProperty.AppEUI] = "012345678901234567890123456789FF";
-            twin.Properties.Desired[TwinProperty.AppKey] = simulatedDevice.LoRaDevice.AppKey;
-            twin.Properties.Desired[TwinProperty.GatewayID] = deviceGatewayID;
-            twin.Properties.Desired[TwinProperty.SensorDecoder] = simulatedDevice.LoRaDevice.SensorDecoder;
-            loRaDeviceClient.Setup(x => x.GetTwinAsync()).ReturnsAsync(twin);
-         
-
-            // Lora device api will be search by devices with matching deveui,
-            var loRaDeviceApi = new Mock<LoRaDeviceAPIServiceBase>(MockBehavior.Strict);
-            loRaDeviceApi.Setup(x => x.SearchDevicesAsync(ServerConfiguration.GatewayID, null, devEUI, appEUI, joinRequestDevNonce))
-                .ReturnsAsync(new SearchDevicesResult(new IoTHubDeviceInfo(devAddr, devEUI, "aabb").AsList()));
-
-            // using factory to create mock of 
-            var loRaDeviceFactory = new TestLoRaDeviceFactory(loRaDeviceClient.Object);
-
-            var memoryCache = new MemoryCache(new MemoryCacheOptions());
-            var deviceRegistry = new LoRaDeviceRegistry(this.ServerConfiguration, memoryCache, loRaDeviceApi.Object, loRaDeviceFactory);
-
-            var frameCounterUpdateStrategyFactory = new LoRaDeviceFrameCounterUpdateStrategyFactory(ServerConfiguration.GatewayID, loRaDeviceApi.Object);
-
-            var messageProcessor = new LoRaWan.NetworkServer.V2.MessageProcessor(
-                this.ServerConfiguration,
-                deviceRegistry,
-                frameCounterUpdateStrategyFactory,
-                new LoRaPayloadDecoder()
-                );
-
-            // join request should fail
-            var joinRequestDownlinkMessage = await messageProcessor.ProcessMessageAsync(joinRequestRxpk);
-            Assert.Null(joinRequestDownlinkMessage);
-
-            loRaDeviceClient.Verify(x => x.UpdateReportedPropertiesAsync(It.IsAny<TwinCollection>()), Times.Never);
-
-            loRaDeviceClient.VerifyAll();
-            loRaDeviceApi.VerifyAll();
-        }
-
-        [Theory]
-        [InlineData(MessageProcessorTestBase.ServerGatewayID)]
-        [InlineData(null)]
-        public async Task Join_Device_Has_Mismatching_AppKey_Should_Return_Null(string deviceGatewayID)
-        {
-            var simulatedDevice = new SimulatedDevice(TestDeviceInfo.CreateOTAADevice(1, gatewayID: deviceGatewayID));
-            var joinRequest = simulatedDevice.CreateJoinRequest();
-
-            // Create Rxpk
-            var joinRequestRxpk = joinRequest.SerializeUplink(simulatedDevice.AppKey).rxpk[0];
-
-            var joinRequestDevNonce = LoRaTools.Utils.ConversionHelper.ByteArrayToString(joinRequest.DevNonce);
-            var devAddr = string.Empty;
-            var devEUI = simulatedDevice.LoRaDevice.DeviceID;
-            var appEUI = simulatedDevice.LoRaDevice.AppEUI;
-
-            var loRaDeviceClient = new Mock<ILoRaDeviceClient>(MockBehavior.Strict);
-
-            // Device twin will be queried
-            var twin = new Twin();
-            twin.Properties.Desired[TwinProperty.DevEUI] = devEUI;
-            twin.Properties.Desired[TwinProperty.AppEUI] = simulatedDevice.LoRaDevice.AppEUI;
-            twin.Properties.Desired[TwinProperty.AppKey] = "012345678901234567890123456789FF";
-            twin.Properties.Desired[TwinProperty.GatewayID] = deviceGatewayID;
-            twin.Properties.Desired[TwinProperty.SensorDecoder] = simulatedDevice.LoRaDevice.SensorDecoder;
-            loRaDeviceClient.Setup(x => x.GetTwinAsync()).ReturnsAsync(twin);
-
-
-            // Lora device api will be search by devices with matching deveui,
-            var loRaDeviceApi = new Mock<LoRaDeviceAPIServiceBase>(MockBehavior.Strict);
-            loRaDeviceApi.Setup(x => x.SearchDevicesAsync(ServerConfiguration.GatewayID, null, devEUI, appEUI, joinRequestDevNonce))
-                .ReturnsAsync(new SearchDevicesResult(new IoTHubDeviceInfo(devAddr, devEUI, "aabb").AsList()));
-
-            // using factory to create mock of 
-            var loRaDeviceFactory = new TestLoRaDeviceFactory(loRaDeviceClient.Object);
-
-            var memoryCache = new MemoryCache(new MemoryCacheOptions());
-            var deviceRegistry = new LoRaDeviceRegistry(this.ServerConfiguration, memoryCache, loRaDeviceApi.Object, loRaDeviceFactory);
-
-            var frameCounterUpdateStrategyFactory = new LoRaDeviceFrameCounterUpdateStrategyFactory(ServerConfiguration.GatewayID, loRaDeviceApi.Object);
-
-            var messageProcessor = new LoRaWan.NetworkServer.V2.MessageProcessor(
-                this.ServerConfiguration,
-                deviceRegistry,
-                frameCounterUpdateStrategyFactory,
-                new LoRaPayloadDecoder()
-                );
-
-            // join request should fail
-            var joinRequestDownlinkMessage = await messageProcessor.ProcessMessageAsync(joinRequestRxpk);
-            Assert.Null(joinRequestDownlinkMessage);
-
-            loRaDeviceClient.Verify(x => x.UpdateReportedPropertiesAsync(It.IsAny<TwinCollection>()), Times.Never);
-            loRaDeviceClient.VerifyAll();
-            loRaDeviceApi.VerifyAll();
-        }
-
         [Theory]
         [InlineData(MessageProcessorTestBase.ServerGatewayID, null)]
         [InlineData(MessageProcessorTestBase.ServerGatewayID, "test")]
@@ -1440,83 +847,6 @@ namespace LoRaWan.NetworkServer.Test
         }
 
 
-        /// <summary>
-        /// Downlink should use same rfch than uplink message
-        /// RFCH stands for Concentrator "RF chain" used for RX
-        /// </summary>
-        /// <param name="deviceGatewayID"></param>
-        /// <param name="rfch"></param>
-        /// <returns></returns>
-        [Theory]
-        [InlineData(MessageProcessorTestBase.ServerGatewayID, 1)]
-        [InlineData(MessageProcessorTestBase.ServerGatewayID, 0)]
-        [InlineData(null, 1)]
-        [InlineData(null, 0)]
-        public async Task OTAA_Join_Should_Use_Rchf_0(string deviceGatewayID, uint rfch)
-        {
-            var simulatedDevice = new SimulatedDevice(TestDeviceInfo.CreateOTAADevice(1, gatewayID: deviceGatewayID));
-            var joinRequest = simulatedDevice.CreateJoinRequest();
-
-            // Create Rxpk
-            var joinRxpk = joinRequest.SerializeUplink(simulatedDevice.LoRaDevice.AppKey).rxpk[0];
-            joinRxpk.rfch = rfch;
-
-            var devNonce = LoRaTools.Utils.ConversionHelper.ByteArrayToString(joinRequest.DevNonce);
-            var devAddr = string.Empty;
-            var devEUI = simulatedDevice.LoRaDevice.DeviceID;
-            var appEUI = simulatedDevice.LoRaDevice.AppEUI;
-
-            var loRaDeviceClient = new Mock<ILoRaDeviceClient>(MockBehavior.Strict);
-
-            // Device twin will be queried
-            var twin = new Twin();
-            twin.Properties.Desired[TwinProperty.DevEUI] = devEUI;
-            twin.Properties.Desired[TwinProperty.AppEUI] = simulatedDevice.LoRaDevice.AppEUI;
-            twin.Properties.Desired[TwinProperty.AppKey] = simulatedDevice.LoRaDevice.AppKey;
-            twin.Properties.Desired[TwinProperty.GatewayID] = deviceGatewayID;
-            twin.Properties.Desired[TwinProperty.SensorDecoder] = simulatedDevice.LoRaDevice.SensorDecoder;
-            loRaDeviceClient.Setup(x => x.GetTwinAsync()).ReturnsAsync(twin);
-
-            // Device twin will be updated
-            loRaDeviceClient.Setup(x => x.UpdateReportedPropertiesAsync(It.IsNotNull<TwinCollection>()))
-                .ReturnsAsync(true);
-
-
-            // Lora device api will be search by devices with matching deveui,
-            var loRaDeviceApi = new Mock<LoRaDeviceAPIServiceBase>(MockBehavior.Strict);
-            loRaDeviceApi.Setup(x => x.SearchDevicesAsync(ServerConfiguration.GatewayID, null, devEUI, appEUI, devNonce))
-                .ReturnsAsync(new SearchDevicesResult(new IoTHubDeviceInfo(devAddr, devEUI, "aabb").AsList()));
-
-            // using factory to create mock of 
-            var loRaDeviceFactory = new TestLoRaDeviceFactory(loRaDeviceClient.Object);
-
-            var memoryCache = new MemoryCache(new MemoryCacheOptions());
-            var deviceRegistry = new LoRaDeviceRegistry(this.ServerConfiguration, memoryCache, loRaDeviceApi.Object, loRaDeviceFactory);
-
-            var frameCounterUpdateStrategyFactory = new LoRaDeviceFrameCounterUpdateStrategyFactory(ServerConfiguration.GatewayID, loRaDeviceApi.Object);
-
-            // Send to message processor
-            var messageProcessor = new LoRaWan.NetworkServer.V2.MessageProcessor(
-                this.ServerConfiguration,
-                deviceRegistry,
-                frameCounterUpdateStrategyFactory,
-                new LoRaPayloadDecoder()
-                );
-
-            var downlinkJoinAcceptMessage = await messageProcessor.ProcessMessageAsync(joinRxpk);
-            Assert.NotNull(downlinkJoinAcceptMessage);
-            // validates txpk according to eu region
-            Assert.Equal((uint)0, downlinkJoinAcceptMessage.txpk.rfch);
-            Assert.Equal(RegionFactory.CreateEU868Region().GetDownstreamChannel(joinRxpk), downlinkJoinAcceptMessage.txpk.freq);
-            Assert.Equal("4/5", downlinkJoinAcceptMessage.txpk.codr);
-            Assert.False(downlinkJoinAcceptMessage.txpk.imme);
-            Assert.True(downlinkJoinAcceptMessage.txpk.ipol);
-            Assert.Equal("LORA", downlinkJoinAcceptMessage.txpk.modu);
-
-            loRaDeviceClient.VerifyAll();
-            loRaDeviceApi.VerifyAll();
-        }
-
         [Theory]
         [InlineData(ServerGatewayID, 1600)]
         [InlineData(ServerGatewayID, 2000)]
@@ -1817,5 +1147,97 @@ namespace LoRaWan.NetworkServer.Test
             loRaDeviceApi.VerifyAll();
 
         }
+
+
+        /// <summary>
+        /// Tests that a ABP device (already in cached or not), receives 1st message with invalid mic, 2nd with valid
+        /// should send message 2 to iot hub
+        /// </summary>
+        /// <returns></returns>
+        [Theory]
+        [InlineData(false)]
+        [InlineData(true)]
+        public async Task ABP_When_First_Message_Has_Invalid_Mic_Second_Should_Send_To_Hub(bool isAlreadyInDeviceRegistryCache)
+        {
+            var simulatedDevice = new SimulatedDevice(TestDeviceInfo.CreateABPDevice(1, gatewayID: ServerGatewayID));
+            var loRaDeviceClient = new Mock<ILoRaDeviceClient>(MockBehavior.Strict);
+            var loRaDevice = TestUtils.CreateFromSimulatedDevice(simulatedDevice, loRaDeviceClient.Object);
+            loRaDevice.SensorDecoder = "DecoderValueSensor";
+
+            const int firstMessageFcnt = 3;
+            const int secondMessageFcnt = 4;
+            const string wrongNwkSKey = "00000000000000000000000000001234";
+            var unconfirmedMessageWithWrongMic = simulatedDevice.CreateUnconfirmedDataUpMessage("123", fcnt: firstMessageFcnt).SerializeUplink(simulatedDevice.AppSKey, wrongNwkSKey).rxpk[0];
+            var unconfirmedMessageWithCorrectMic = simulatedDevice.CreateUnconfirmedDataUpMessage("456", fcnt: secondMessageFcnt).SerializeUplink(simulatedDevice.AppSKey, simulatedDevice.NwkSKey).rxpk[0];
+
+            // message will be sent
+            LoRaDeviceTelemetry loRaDeviceTelemetry = null;
+            loRaDeviceClient.Setup(x => x.SendEventAsync(It.IsNotNull<LoRaDeviceTelemetry>(), null))
+                .Callback<LoRaDeviceTelemetry, Dictionary<string, string>>((t, _) => loRaDeviceTelemetry = t)
+                .ReturnsAsync(true);
+
+            if (!isAlreadyInDeviceRegistryCache)
+            {
+                loRaDeviceClient.Setup(x => x.GetTwinAsync())
+                    .ReturnsAsync(simulatedDevice.CreateABPTwin());
+            }
+
+
+            // C2D message will be checked
+            loRaDeviceClient.Setup(x => x.ReceiveAsync(It.IsNotNull<TimeSpan>()))
+                .ReturnsAsync((Message)null);
+
+            // Lora device api
+            var loRaDeviceApi = new Mock<LoRaDeviceAPIServiceBase>(MockBehavior.Strict);
+
+            // will search for the device twice
+            loRaDeviceApi.Setup(x => x.SearchDevicesAsync(null, loRaDevice.DevAddr, null, null, null))
+                .ReturnsAsync(new SearchDevicesResult(new IoTHubDeviceInfo(loRaDevice.DevAddr, loRaDevice.DevEUI, "aaa").AsList()));
+
+            // using factory to create mock of 
+            var loRaDeviceFactory = new TestLoRaDeviceFactory(loRaDeviceClient.Object);
+
+            // add device to cache already
+            var memoryCache = new MemoryCache(new MemoryCacheOptions());
+            if (isAlreadyInDeviceRegistryCache)
+            {
+                var dictionary = new DevEUIToLoRaDeviceDictionary();
+                dictionary[loRaDevice.DevEUI] = loRaDevice;
+                memoryCache.Set<DevEUIToLoRaDeviceDictionary>(loRaDevice.DevAddr, dictionary);
+            }
+
+            var deviceRegistry = new LoRaDeviceRegistry(this.ServerConfiguration, memoryCache, loRaDeviceApi.Object, loRaDeviceFactory);
+
+            var frameCounterUpdateStrategyFactory = new LoRaDeviceFrameCounterUpdateStrategyFactory(ServerConfiguration.GatewayID, loRaDeviceApi.Object);
+
+            // Send to message processor
+            var messageProcessor = new LoRaWan.NetworkServer.V2.MessageProcessor(
+                this.ServerConfiguration,
+                deviceRegistry,
+                frameCounterUpdateStrategyFactory,
+                new LoRaPayloadDecoder()
+                );
+
+            // first message should fail
+            Assert.Null(await messageProcessor.ProcessMessageAsync(unconfirmedMessageWithWrongMic));
+
+            // second message should succeed
+            Assert.Null(await messageProcessor.ProcessMessageAsync(unconfirmedMessageWithCorrectMic));
+
+
+            Assert.NotNull(loRaDeviceTelemetry);
+            Assert.IsType<JObject>(loRaDeviceTelemetry.data);
+            var telemetryData = (JObject)loRaDeviceTelemetry.data;
+            Assert.Equal("456", telemetryData["value"].ToString());
+
+            var devicesByDevAddr = deviceRegistry.InternalGetCachedDevicesForDevAddr(simulatedDevice.DevAddr);
+            Assert.NotEmpty(devicesByDevAddr);
+            Assert.True(devicesByDevAddr.TryGetValue(simulatedDevice.DevEUI, out var loRaDeviceFromRegistry));
+            Assert.Equal(secondMessageFcnt, loRaDeviceFromRegistry.FCntUp);
+            Assert.True(loRaDeviceFromRegistry.IsOurDevice);
+
+            loRaDeviceApi.VerifyAll();
+            loRaDeviceClient.VerifyAll();
+        }       
     }
 }
